Move HabitStackScreen into stack/
