+++ features/habit/HabitListScreen.tsx
@@ -2,7 +2,10 @@ import { StackNavigationProp } from "@react-navigation/stack";
 import React, { FC } from "react";
 
 import { HabitList } from "./HabitList";
-import { HabitStackParamList, HabitStackScreenName } from "./HabitStackScreen";
+import {
+  HabitStackParamList,
+  HabitStackScreenName,
+} from "./stack/HabitStackScreen";
 
 type HabitListScreenNavigationProp = StackNavigationProp<
   HabitStackParamList,
